refactor(env): Extract `util` into separate file
